Add licher, fixed internal error
Original Code by t.me/TheKneesocks

+++ lazyleech/utils/upload_worker.py
@@ -15,6 +15,7 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import os
+import re
 import html
 import time
 import shutil
@@ -23,14 +24,17 @@ import asyncio
 import zipfile
 import tempfile
 import traceback
+from collections import defaultdict
 from natsort import natsorted
 from pyrogram.parser import html as pyrogram_html
-from .. import PROGRESS_UPDATE_DELAY, ADMIN_CHATS, preserved_logs, TESTMODE, SendAsZipFlag, ForceDocumentFlag
+from .. import PROGRESS_UPDATE_DELAY, ADMIN_CHATS, ALL_CHATS, preserved_logs, TESTMODE, SendAsZipFlag, ForceDocumentFlag, LICHER_CHAT, LICHER_STICKER, LICHER_FOOTER, LICHER_PARSE_EPISODE
 from .misc import split_files, get_file_mimetype, format_bytes, get_video_info, generate_thumbnail, return_progress_string, calculate_eta, watermark_photo
 
 upload_queue = asyncio.Queue()
 upload_statuses = dict()
 upload_tamper_lock = asyncio.Lock()
+message_exists = defaultdict(set)
+message_exists_lock = asyncio.Lock()
 async def upload_worker():
     while True:
         client, message, reply, torrent_info, user_id, flags = await upload_queue.get()
@@ -91,11 +95,14 @@ async def _upload_worker(client, message, reply, torrent_info, user_id, flags):
         else:
             for file in torrent_info['files']:
                 filepath = file['path']
-                filename = filepath.replace(os.path.join(torrent_info['dir'], ''), '', 1)
+                if LICHER_PARSE_EPISODE:
+                    filename = re.sub(r'\s*(?:\[.+?\]|\(.+?\))\s*|\.[a-z][a-z0-9]{2}$', '', os.path.basename(filepath))
+                else:
+                    filename = filepath.replace(os.path.join(torrent_info['dir'], ''), '', 1)
                 files[filepath] = filename
         for filepath in natsorted(files):
             sent_files.extend(await _upload_file(client, message, reply, files[filepath], filepath, ForceDocumentFlag in flags))
-    text = '✨ #Files:\n'
+    text = '✨ <b>#Files:</b> \n\n'
     parser = pyrogram_html.HTML(client)
     quote = None
     first_index = None
@@ -119,6 +126,8 @@ async def _upload_worker(client, message, reply, torrent_info, user_id, flags):
         text = futtext
     if not sent_files:
         text = 'Files: None'
+    elif LICHER_CHAT and LICHER_STICKER and message.chat.id in ADMIN_CHATS:
+        await client.send_sticker(LICHER_CHAT, LICHER_STICKER)
     thing = await message.reply_text(text, quote=quote, disable_web_page_preview=True)
     if first_index is None:
         first_index = thing
@@ -134,6 +143,7 @@ async def _upload_file(client, message, reply, filename, filepath, force_documen
     user_watermarked_thumbnail = os.path.join(str(user_id), 'watermarked_thumbnail.jpg')
     file_has_big = os.path.getsize(filepath) > 2097152000
     upload_wait = await reply.reply_text(f'Upload of {html.escape(filename)} will start in {PROGRESS_UPDATE_DELAY}s')
+    message_exists[upload_wait.chat.id].add(upload_wait.message_id)
     upload_identifier = (upload_wait.chat.id, upload_wait.message_id)
     async with upload_tamper_lock:
         upload_waits[upload_identifier] = user_id, worker_identifier
@@ -182,7 +192,7 @@ async def _upload_file(client, message, reply, filename, filepath, force_documen
                     for i in (user_thumbnail, user_watermarked_thumbnail):
                         thumbnail = i if os.path.isfile(i) else thumbnail
                     mimetype = await get_file_mimetype(filepath)
-                    progress_args = (client, upload_wait, filename, user_id)
+                    progress_args = (client, message, upload_wait, filename, user_id)
                     try:
                         if not force_document and mimetype.startswith('video/'):
                             duration = 0
@@ -222,38 +232,56 @@ async def _upload_file(client, message, reply, filename, filepath, force_documen
                         continue
                     if resp:
                         sent_files.append((os.path.basename(filename), resp.link))
+                        if LICHER_CHAT and reply.chat.id in ALL_CHATS and mimetype.startswith('video/') and resp.video:
+                            await client.send_video(LICHER_CHAT, resp.video.file_id, thumb=thumbnail,
+                                                    caption=filename + LICHER_FOOTER, duration=duration,
+                                                    width=width, height=height, parse_mode=None)
                         break
                     return sent_files
         return sent_files
     finally:
         if split_task:
             split_task.cancel()
+        async with message_exists_lock:
+            message_exists[upload_wait.chat.id].discard(upload_wait.message_id)
         asyncio.create_task(upload_wait.delete())
         async with upload_tamper_lock:
             upload_waits.pop(upload_identifier)
 
 progress_callback_data = dict()
 stop_uploads = set()
-async def progress_callback(current, total, client, reply, filename, user_id):
-    message_identifier = (reply.chat.id, reply.message_id)
-    last_edit_time, prevtext, start_time, user_id = progress_callback_data.get(message_identifier, (0, None, time.time(), user_id))
-    if message_identifier in stop_uploads or current == total:
-        asyncio.create_task(reply.delete())
-        try:
-            progress_callback_data.pop(message_identifier)
-        except KeyError:
-            pass
-        if message_identifier in stop_uploads:
-            client.stop_transmission()
-    elif (time.time() - last_edit_time) > PROGRESS_UPDATE_DELAY:
-        if last_edit_time:
-            upload_speed = format_bytes((total - current) / (time.time() - start_time))
-        else:
-            upload_speed = '0 B'
-        text = f'''<b>Uploading</b> <code>{html.escape(filename)}</code>
+async def progress_callback(current, total, client, message, reply, filename, user_id):
+    try:
+        if reply.message_id not in message_exists[reply.chat.id]:
+            return
+        message_identifier = (reply.chat.id, reply.message_id)
+        last_edit_time, prevtext, start_time, user_id = progress_callback_data.get(message_identifier, (0, None, time.time(), user_id))
+        if message_identifier in stop_uploads or current == total:
+            asyncio.create_task(reply.delete())
+            try:
+                progress_callback_data.pop(message_identifier)
+            except KeyError:
+                pass
+            if message_identifier in stop_uploads:
+                client.stop_transmission()
+        elif (time.time() - last_edit_time) > PROGRESS_UPDATE_DELAY:
+            if last_edit_time:
+                upload_speed = format_bytes((total - current) / (time.time() - start_time))
+            else:
+                upload_speed = '0 B'
+            text = f'''<b>Uploading</b> <code>{html.escape(filename)}</code>
 <code>{html.escape(return_progress_string(current, total))}</code> {format_bytes(current)} of {format_bytes(total)} at {upload_speed}/s, ETA: {calculate_eta(current, total, start_time)}'''
-        if prevtext != text:
-            await reply.edit_text(text)
-            prevtext = text
-            last_edit_time = time.time()
-            progress_callback_data[message_identifier] = last_edit_time, prevtext, start_time, user_id
+            if prevtext != text and reply.message_id in message_exists[reply.chat.id]:
+                async with message_exists_lock:
+                    if reply.message_id not in message_exists[reply.chat.id]:
+                        return
+                    await reply.edit_text(text)
+                prevtext = text
+                last_edit_time = time.time()
+                progress_callback_data[message_identifier] = last_edit_time, prevtext, start_time, user_id
+    except Exception as ex:
+        preserved_logs.append((message, None, ex))
+        logging.exception('%s', message)
+        await message.reply_text(traceback.format_exc(), parse_mode=None)
+        for admin_chat in ADMIN_CHATS:
+            await client.send_message(admin_chat, traceback.format_exc(), parse_mode=None)
